+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Items that are related, such as breaking changes, new features, or changes to ex
 
 - [@siteimprove/alfa-cypress](packages/alfa-cypress): `Cypress.createPlugin()` no longer creates an asynchronous Chai plugin, which was causing assertion errors to not count as test failures.
 
+- [@siteimprove/alfa-rules](packages/alfa-rules): SIA-R84 now only applies to elements that are possibly scrollable along the x-axis if they also specify `white-space: nowrap`.
+
 ## [0.15.0](../../compare/v0.14.2...v0.15.0) (2021-03-15)
 
 ### Breaking