+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Items that are related, such as breaking changes, new features, or changes to ex
 
 ## [Unreleased]
 
+## [0.15.1](../../compare/v0.15.0...v0.15.1) (2021-03-17)
+
 ### Fixed
 
 - [@siteimprove/alfa-rules](packages/alfa-rules): `isClipped()` now correctly handles text nodes of clipped elements.